Pass uint16_t for window extents
The underlying bitmap layer operates in uint16_t, so use it at this
layer also, to avoid narrowing when forwarding to the bitmap code.

+++ common/include/window.h
@@ -38,7 +38,7 @@ private:
 	uint8_t w_visible = 1;						// whether it's visible
 	uint8_t w_modal = 1;						// modal = accept all user input exclusively
 public:
-	explicit window(grs_canvas &src, int x, int y, int w, int h);
+	explicit window(grs_canvas &src, uint16_t x, uint16_t y, uint16_t w, uint16_t h);
 	window(const window &) = delete;
 	window &operator=(const window &) = delete;
 	virtual ~window();
